+++ Scripts/Items/Temrael/Meubles/ArmesEntrainement.cs
@@ -5,8 +5,8 @@ namespace Server.Items
     public class DagueBois : BaseKnife
 	{
         public override int DefSpeed { get { return 20; } }
-        public override int DefMinDamage { get { return 1; } }
-        public override int DefMaxDamage { get { return 2; } }
+        public override double DefMinDamage { get { return 1; } }
+        public override double DefMaxDamage { get { return 2; } }
         
 		[Constructable]
 		public DagueBois() : base( 0x1494 )
@@ -39,8 +39,8 @@ namespace Server.Items
     public class LanceBois : BaseSpear
     {
         public override int DefSpeed { get { return 20; } }
-        public override int DefMinDamage { get { return 1; } }
-        public override int DefMaxDamage { get { return 2; } }
+        public override double DefMinDamage { get { return 1; } }
+        public override double DefMaxDamage { get { return 2; } }
 
         [Constructable]
         public LanceBois()
@@ -74,8 +74,8 @@ namespace Server.Items
     public class MasseBois : BaseBashing
     {
         public override int DefSpeed { get { return 20; } }
-        public override int DefMinDamage { get { return 1; } }
-        public override int DefMaxDamage { get { return 2; } }
+        public override double DefMinDamage { get { return 1; } }
+        public override double DefMaxDamage { get { return 2; } }
 
         [Constructable]
         public MasseBois()
@@ -109,8 +109,8 @@ namespace Server.Items
     public class BatonBois : BaseStaff
     {
         public override int DefSpeed { get { return 20; } }
-        public override int DefMinDamage { get { return 1; } }
-        public override int DefMaxDamage { get { return 2; } }
+        public override double DefMinDamage { get { return 1; } }
+        public override double DefMaxDamage { get { return 2; } }
 
         [Constructable]
         public BatonBois()
@@ -144,8 +144,8 @@ namespace Server.Items
     public class EpeeBois : BaseSword
     {
         public override int DefSpeed { get { return 20; } }
-        public override int DefMinDamage { get { return 1; } }
-        public override int DefMaxDamage { get { return 2; } }
+        public override double DefMinDamage { get { return 1; } }
+        public override double DefMaxDamage { get { return 2; } }
 
         [Constructable]
         public EpeeBois()
